+++ CHANGELOG.md
@@ -1,3 +1,17 @@
 ## 0.0.1
 
 * initial release.
+
+## 0.0.2
+
+* Format code.
+
+* Modified app store channel name
+
+## 0.0.3
+
+* Add ios kill app
+
+## 0.0.4
+
+* Migrate to null-safety